requirements.txt
Flask>=2.3.2
Flask-Cors>=4.0.0
openai>=1.0.0
gunicorn
Flask-Session>=0.5.0
firebase-admin>=6.2.0
python-dotenv>=1.0.0